requirements.txt
langchain
ipykernel
openai
fastapi
langchain_groq
langchain_core
langchain_community
langserve
python-dotenv
streamlit
transformers
langchain_openai
streamlit
pypdf
google-generativeai
langchain_huggingface
sentence-transformers
langchain_google_genai
arxiv
wikipedia
duckdb
duckduckgo-search
validators
unstructured
nltk
youtube-transcript-api
pytube
numexpr
cassio
datasets
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools
neo4j==5.14
langchain_experimental
lamini
python-docx